Make the ABI comparison a Package QA test
The abicheck class currently hooks two functions after the do_install
step: one to gather the ABI data for this build, and one to perform the
ABI comparison if a baseline is available.

For the second, we now use Package QA instead to do the ABI comparison,
allowing a user to specify whether if they want an ABI change to be a
warning or an error.